pyproject.toml
[tool.poetry]
name = "miningniti"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.350"
sentence-transformers = "^2.2.2"
pypdf2 = "^3.0.1"
accelerate = "^0.25.0"
pyyaml = "^6.0.1"
fastapi = "^0.105.0"
uvicorn = "^0.24.0.post1"
faiss-cpu = "^1.7.4" # use faiss for cpu, we've enough load on GPU already

bitsandbytes = {path = "libs/bitsandbytes-0.41.1-py3-none-win_amd64.whl"}
torch = "^2.1.0+cu121"
torchvision = "^0.16.0+cu121"
torchaudio = "^2.1.0+cu121"

# https://github.com/jllllll/bitsandbytes-windows-webui/releases/download/wheels/bitsandbytes-0.41.1-py3-none-win_amd64.whl
# pip install torch==2.1.0 torchvision==0.16.0 torchaudio==2.1.0 --index-url https://download.pytorch.org/whl/cu121

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"